[freechart] Bugfix: Vertical axis title extent calculation error
Extents were being miscalculated for vertical axes titles. This left a lot of white space to the right side of the title. Basically, the longer the label string, the more space was left. The root cause of this was the calculations being based on the height of the label text for horizontal label strings, and the width for vertical strings...