[freechart] BUGFIX: Plotting zero values on log charts causes draw errors
When the incoming data series for a given axis contains zero values, the LogarithmicNumberAxis will fail to plot any data. Axis value labels are not shown, nor is any data actually plotted.

This is caused by inadequate error-recovery (or even error trapping) in the LogNumAxis code. This patch adds defenses to the code which will cause values less than zero to be plotted as zero.

Ideally the log-axis should implement skewing (shift the input to log(n) to keep it in the range 1 <= n <= infinity) but this will work as an interim patch. See http://www.purplemath.com/modules/graphlog3.htm for more about this.